# core-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Make-Singer — AI-powered music creation platform.",
        "title": "Make-Singer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/render/cancel/{jobId}": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Cancel a running or queued render job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Render"
                ],
                "summary": "Cancel render job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RenderCancelResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/render/result/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the result of a completed render job including stems",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Render"
                ],
                "summary": "Get render job result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RenderResultResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/render/start": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Start an asynchronous render job for a song arrangement",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Render"
                ],
                "summary": "Start render job",
                "parameters": [
                    {
                        "description": "Render start request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.RenderStartRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/model.RenderStartResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/render/status/{jobId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the current status and progress of a render job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Render"
                ],
                "summary": "Get render job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "jobId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.RenderStatusResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/upload/vocal": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Upload an audio file as a vocal take for a song section",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Upload vocal take",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Section ID",
                        "name": "sectionId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Take name",
                        "name": "takeName",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Audio file (WAV, MP3, M4A, AAC; max 50MB)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.UploadVocalResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Too Many Requests",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/upload/vocal/{takeId}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a previously uploaded vocal take",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Upload"
                ],
                "summary": "Delete vocal take",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Take ID",
                        "name": "takeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Arrangement": {
            "type": "object",
            "required": [
                "density",
                "groove",
                "instruments"
            ],
            "properties": {
                "density": {
                    "enum": [
                        "minimal",
                        "medium",
                        "full"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Density"
                        }
                    ]
                },
                "groove": {
                    "enum": [
                        "straight",
                        "swing",
                        "half_time"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Groove"
                        }
                    ]
                },
                "instruments": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.Instrument"
                    }
                },
                "sectionEmphasis": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.SectionEmphasis"
                    }
                }
            }
        },
        "model.BPMConfig": {
            "type": "object",
            "required": [
                "mode"
            ],
            "properties": {
                "max": {
                    "type": "integer",
                    "maximum": 220,
                    "minimum": 40
                },
                "min": {
                    "type": "integer",
                    "maximum": 220,
                    "minimum": 40
                },
                "mode": {
                    "enum": [
                        "auto",
                        "range",
                        "fixed"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.BPMMode"
                        }
                    ]
                },
                "value": {
                    "type": "integer",
                    "maximum": 220,
                    "minimum": 40
                }
            }
        },
        "model.BPMMode": {
            "type": "string",
            "enum": [
                "auto",
                "range",
                "fixed"
            ],
            "x-enum-varnames": [
                "BPMModeAuto",
                "BPMModeRange",
                "BPMModeFixed"
            ]
        },
        "model.Brief": {
            "type": "object",
            "required": [
                "bpm",
                "genre",
                "key",
                "structure",
                "vibes"
            ],
            "properties": {
                "bpm": {
                    "$ref": "#/definitions/model.BPMConfig"
                },
                "genre": {
                    "enum": [
                        "pop",
                        "rock",
                        "hiphop",
                        "rnb",
                        "electronic",
                        "jazz",
                        "country",
                        "folk",
                        "classical",
                        "latin",
                        "reggae",
                        "blues"
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.Genre"
                        }
                    ]
                },
                "key": {
                    "$ref": "#/definitions/model.KeyConfig"
                },
                "structure": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/model.SongSection"
                    }
                },
                "vibes": {
                    "type": "array",
                    "maxItems": 5,
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "model.Density": {
            "type": "string",
            "enum": [
                "minimal",
                "medium",
                "full"
            ],
            "x-enum-varnames": [
                "DensityMinimal",
                "DensityMedium",
                "DensityFull"
            ]
        },
...
        "model.Groove": {
            "type": "string",
            "enum": [
                "straight",
                "swing",
                "half_time"
            ],
            "x-enum-varnames": [
                "GrooveStraight",
                "GrooveSwing",
                "GrooveHalfTime"
            ]
        },
        "model.Instrument": {
            "type": "string",
            "enum": [
                "drums",
                "bass",
                "piano",
                "guitar",
                "synth",
                "strings",
                "brass",
                "woodwinds",
                "percussion",
                "pads",
                "lead",
                "fx"
            ],
            "x-enum-varnames": [
                "InstrumentDrums",
                "InstrumentBass",
                "InstrumentPiano",
                "InstrumentGuitar",
                "InstrumentSynth",
                "InstrumentStrings",
                "InstrumentBrass",
                "InstrumentWoodwinds",
                "InstrumentPercussion",
                "InstrumentPads",
                "InstrumentLead",
                "InstrumentFX"
            ]
        },
        "model.JobStatus": {
            "type": "string",
            "enum": [
                "queued",
                "running",
                "succeeded",
                "failed",
                "canceled"
            ],
            "x-enum-varnames": [
                "JobStatusQueued",
                "JobStatusRunning",
                "JobStatusSucceeded",
                "JobStatusFailed",
                "JobStatusCanceled"
            ]
        },
...
        "model.KeyResult": {
            "type": "object",
            "properties": {
                "scale": {
                    "$ref": "#/definitions/model.Scale"
                },
                "tonic": {
                    "$ref": "#/definitions/model.Tonic"
                }
            }
        },
...
        "model.RenderCancelResponse": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "model.RenderResultResponse": {
            "type": "object",
            "properties": {
                "bpm": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "key": {
                    "$ref": "#/definitions/model.KeyResult"
                },
                "stems": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.StemResult"
                    }
                }
            }
        },
        "model.RenderStartRequest": {
            "type": "object",
            "required": [
                "arrangement",
                "brief",
                "projectId"
            ],
            "properties": {
                "arrangement": {
                    "$ref": "#/definitions/model.Arrangement"
                },
                "brief": {
                    "$ref": "#/definitions/model.Brief"
                },
                "projectId": {
                    "type": "string"
                }
            }
        },
        "model.RenderStartResponse": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "estimatedDuration": {
                    "type": "integer"
                },
                "jobId": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                }
            }
        },
        "model.RenderStatusResponse": {
            "type": "object",
            "properties": {
                "completedAt": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentStep": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                },
                "jobId": {
                    "type": "string"
                },
                "progress": {
                    "type": "integer"
                },
                "retryCount": {
                    "type": "integer"
                },
                "startedAt": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.JobStatus"
                }
            }
        },
        "model.Scale": {
            "type": "string",
            "enum": [
                "major",
                "minor"
            ],
            "x-enum-varnames": [
                "ScaleMajor",
                "ScaleMinor"
            ]
        },
        "model.SectionEmphasis": {
            "type": "object",
            "required": [
                "emphasis",
                "sectionId"
            ],
            "properties": {
                "emphasis": {
                    "type": "string",
                    "enum": [
                        "bigger",
                        "biggest"
                    ]
                },
                "sectionId": {
                    "type": "string"
                }
            }
        },
...
        "model.StemResult": {
            "type": "object",
            "properties": {
                "duration": {
                    "type": "number"
                },
                "fileUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instrument": {
                    "$ref": "#/definitions/model.Instrument"
                },
                "waveformData": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
        "model.Tonic": {
            "type": "string",
            "enum": [
                "C",
                "C#",
                "D",
                "D#",
                "E",
                "F",
                "F#",
                "G",
                "G#",
                "A",
                "A#",
                "B"
            ],
            "x-enum-varnames": [
                "TonicC",
                "TonicCSharp",
                "TonicD",
                "TonicDSharp",
                "TonicE",
                "TonicF",
                "TonicFSharp",
                "TonicG",
                "TonicGSharp",
                "TonicA",
                "TonicASharp",
                "TonicB"
            ]
        },
        "model.UploadVocalResponse": {
            "type": "object",
            "properties": {
                "channels": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "duration": {
                    "type": "number"
                },
                "fileUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "sampleRate": {
                    "type": "integer"
                }
            }
        },
...
        "response.ErrorDetail": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "message": {
                    "type": "string"
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/response.ErrorDetail"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your bearer token in the format **Bearer \u0026lt;token\u0026gt;**",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}